requirements.txt
Django~=4.1.1
Pillow~=9.2.0
psycopg2~=2.9.5